Update authentication.md
The previous text was either not updated or did not work at any point. id_token no longer seems to be populated in response.params.

Also, the provider no longer needs to be initialized, you can just access it like I did in the code. In fact, the old method is broken.